Fixed Environment.shader optimization.

+++ Assets/Scripts/Shaders/Environment.shader
@@ -8,7 +8,7 @@ Shader "Custom/BackgroundShader" {
 		//Don't mess with this variable the values of these variables will be computed in a script.
 		//_Hue("Hue", Range(0, 360)) = 0
 		_GreenFilter("GreenFilter", Range(0,1)) = 0.25
-		_Hue("Hue", half) = 360
+		_Hue("Hue", Float) = 360
 		_Saturation("Saturation", Range(0,1)) = 0.5820895
 		_TimeFactor("InterpolationFactor", Range(0,1)) = 0
 	}
